refactor to boto authenticator class
