+++ tests/test_aws_authenticators.py
@@ -1,21 +1,22 @@
 import pytest
-from contextlib import nullcontext as does_not_raise
-from unittest import mock
 from moto import mock_dynamodb, mock_sts
-from tap_dynamodb.aws_auth import AWSAuth
-import os
+from tap_dynamodb.aws_authenticators import AWSBotoAuthenticator
 from unittest.mock import patch
-import botocore
 
-@patch("tap_dynamodb.aws_auth.boto3.Session", return_value="mock_session")
+
+@patch(
+    "tap_dynamodb.aws_authenticators.boto3.Session",
+    return_value="mock_session",
+)
 @mock_dynamodb
 def test_get_session_base(patch):
-    auth = AWSAuth(
+    auth = AWSBotoAuthenticator(
         {
             "aws_access_key_id": "foo",
             "aws_secret_access_key": "bar",
             "aws_default_region": "baz",
-        }
+        },
+        "dynamodb",
     )
     session = auth.get_session()
     patch.assert_called_with(
@@ -25,16 +26,18 @@ def test_get_session_base(patch):
     )
     assert session == "mock_session"
 
-@patch("tap_dynamodb.aws_auth.boto3.Session", return_value="mock_session")
+
+@patch("tap_dynamodb.aws_authenticators.boto3.Session", return_value="mock_session")
 @mock_dynamodb
 def test_get_session_w_token(patch):
-    auth = AWSAuth(
+    auth = AWSBotoAuthenticator(
         {
             "aws_access_key_id": "foo",
             "aws_secret_access_key": "bar",
             "aws_session_token": "abc",
             "aws_default_region": "baz",
-        }
+        },
+        "dynamodb",
     )
     session = auth.get_session()
     patch.assert_called_with(
@@ -46,13 +49,14 @@ def test_get_session_w_token(patch):
     assert session == "mock_session"
 
 
-@patch("tap_dynamodb.aws_auth.boto3.Session", return_value="mock_session")
+@patch("tap_dynamodb.aws_authenticators.boto3.Session", return_value="mock_session")
 @mock_dynamodb
 def test_get_session_w_profile(patch):
-    auth = AWSAuth(
+    auth = AWSBotoAuthenticator(
         {
             "aws_profile": "foo",
-        }
+        },
+        "dynamodb",
     )
     session = auth.get_session()
     patch.assert_called_with(
@@ -64,44 +68,48 @@ def test_get_session_w_profile(patch):
 @mock_dynamodb
 def test_get_session_empty_fail():
     with pytest.raises(Exception):
-        auth = AWSAuth({})
+        auth = AWSBotoAuthenticator({})
         auth.get_session()
 
 
 @mock_dynamodb
 @mock_sts
 def test_get_session_assume_role():
-    auth = AWSAuth(
+    auth = AWSBotoAuthenticator(
         {
             "aws_access_key_id": "foo",
             "aws_secret_access_key": "bar",
             "aws_default_region": "baz",
-            "aws_assume_role_arn": "arn:aws:iam::123456778910:role/my-role-name"
-        }
+            "aws_assume_role_arn": "arn:aws:iam::123456778910:role/my-role-name",
+        },
+        "dynamodb",
     )
     session = auth.get_session()
 
 
 @mock_dynamodb
 def test_get_client():
-    auth = AWSAuth(
+    auth = AWSBotoAuthenticator(
         {
             "aws_access_key_id": "foo",
             "aws_secret_access_key": "bar",
             "aws_default_region": "baz",
-        }
+        },
+        "dynamodb",
     )
     session = auth.get_session()
     client = auth.get_client(session, "dynamodb")
 
+
 @mock_dynamodb
 def test_get_resource():
-    auth = AWSAuth(
+    auth = AWSBotoAuthenticator(
         {
             "aws_access_key_id": "foo",
             "aws_secret_access_key": "bar",
             "aws_default_region": "baz",
-        }
+        },
+        "dynamodb",
     )
     session = auth.get_session()
     resource = auth.get_resource(session, "dynamodb")
